+++ mirrord/layer/src/lib.rs
@@ -570,7 +570,7 @@ fn enable_hooks(state: &LayerSetup) {
 #[mirrord_layer_macro::instrument(level = "trace", fields(pid = std::process::id()))]
 pub(crate) fn close_layer_fd(fd: c_int) {
     // Remove from sockets.
-    if let Some((fd, socket)) = SOCKETS.remove(&fd) {
+    if let Some((_, socket)) = SOCKETS.remove(&fd) {
         // Closed file is a socket, so if it's already bound to a port - notify agent to stop
         // mirroring/stealing that port.
         socket.close();
